Refactor subscriber api and add auth guard on get request
