+++ utils/urls.js
@@ -38,9 +38,6 @@ export default {
       index: "/api/blogs",
       publish: "/api/blogs/publish",
     },
-    subscribers: {
-      getAll: "/api/subscribers/getSubscribers",
-      add: "/api/subscribers/addSubscriber",
-    },
+    subscribers: "/api/subscribers",
   },
 };
